fixes for paleo (#9811) and mapping-table issues in converter.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@11702 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/conversion/IdHashMapper.java
@@ -195,14 +195,14 @@ public class IdHashMapper implements IdMapperIFace
         
         try
         {
-            if (enableRemoveRecords)
+            if (enableRemoveRecords && BasicSQLUtils.doesTableExist(oldConn, mapTableName) )
             {
                 wasEmpty = true;
                 try
                 {
                     String sql = String.format("DELETE FROM %s", mapTableName);
                     BasicSQLUtils.update(oldConn, sql);
-                } catch (Exception ex ) {}
+                 } catch (Exception ex ) {}
                 return;
             }
